+++ lib/media/document.ts
@@ -74,7 +74,7 @@ const setupCallback = function (clientObject: Client, evt: AudioEvent): void {
           }
         });
       Plivo.log.debug(
-        `audioDevDictionary is updated onMediaPermission: ${evt.status}`,
+        `${LOGCAT.LOGIN} UserMedia:audioDevDictionary is updated, onMediaPermission:${evt.status}`,
       );
     });
   }
@@ -92,7 +92,6 @@ const setupCallback = function (clientObject: Client, evt: AudioEvent): void {
   clientObject.connectToneView = document.getElementById(
     CONNECT_TONE_ELEMENT_ID,
   ) as HTMLAudioElement;
-  Plivo.log.info(JSON.stringify(clientObject.browserDetails));
 };
 
 /**
@@ -111,15 +110,15 @@ const getLocalMedia = function (
       .getUserMedia({ audio: audioConstraints, video: false })
       .then((stream) => {
         (window as any).localStream = stream;
-        Plivo.log.debug('getUserMedia success');
+        Plivo.log.debug(`${LOGCAT.LOGIN} UserMedia:Access granted for audio`);
         cb(clientObject, { status: 'success', stream: true });
       })
       .catch((err) => {
-        Plivo.log.error(`failed to get user media :: ${err.name}`);
+        Plivo.log.error(`${LOGCAT.LOGIN}UserMedia:Access to audio failed - ${err.name}`);
         cb(clientObject, { status: 'failure', error: err.name });
       });
   } else {
-    Plivo.log.error('getUserMedia not available');
+    Plivo.log.error(`${LOGCAT.LOGIN} UserMedia:Not available`);
     cb(clientObject, {
       status: 'failure',
       error: 'getUserMedia not supported',