+++ test/integration/outgoingcall.js
@@ -137,6 +137,7 @@ describe("plivoWebSdk", function () {
       }, TIMEOUT);
     });
 
+
     // // #16
     // // eslint-disable-next-line no-undef
     // it("outbound call should ring", (done) => {